+++ src/popup/App.tsx
@@ -2,14 +2,14 @@ import { getAllRouter as AllRouter } from "./router";
 import { useIdleTimer } from "react-idle-timer";
 import { sendMsg } from "../utils/commonMsg";
 import { WALLET_RESET_LAST_ACTIVE_TIME } from "../constant/msgTypes";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect } from "react";
 import {
   fetchSupportTokenInfo,
   getRecommendFee,
   getScamList,
 } from "../background/api";
 import { GlobalStyles } from "./GlobalStyles";
-import { StyledApp, StyledAppHeader } from "./App.styled";
+import { StyledApp } from "./App.styled";
 import { updateRecommendFee } from "../reducers/cache";
 import { useDispatch, useSelector } from "react-redux";
 import { getLocal } from "../background/localStorage";
@@ -39,35 +39,6 @@ function App() {
   const dispatch = useDispatch();
   const currentNode = useSelector((state: RootState) => state.network.currentNode);
   const shouldRefresh = useSelector((state: RootState) => state.accountInfo.shouldRefresh);
-
-  const [showFullStatus, setShowFullStatus] = useState(false);
-  const [autoWidthStatus, setAutoWidthStatus] = useState(false);
-  useEffect(() => {
-    const url = new URL(window.location.href);
-    // Extract hash path without query parameters (e.g., "#/register_page?addWallet=true" -> "/register_page")
-    const hashPath = url.hash.split("?")[0]?.replace("#", "") ?? "";
-
-    const fullPageRoutes = ["/ledger_page", "/register_page", "/createprocess"];
-    const zkPageRoutes = ["/approve_page", "/request_sign", "/token_sign"];
-
-    const isZkPage = zkPageRoutes.some((route) => hashPath.startsWith(route));
-    const isFullPage = fullPageRoutes.some((route) =>
-      hashPath.startsWith(route)
-    );
-
-    if (isZkPage) {
-      setAutoWidthStatus(true);
-    }
-
-    if (
-      (url.pathname.indexOf("popup.html") !== -1 && !isFullPage) ||
-      url.pathname.indexOf("notification.html") !== -1
-    ) {
-      setShowFullStatus(false);
-    } else {
-      setShowFullStatus(true);
-    }
-  }, [window.location.href]);
 
   const fetchFeeData = useCallback(async () => {
     const feeConfig = await getRecommendFee();
@@ -141,9 +112,7 @@ function App() {
     <StyledApp>
       <GlobalStyles />
       <LedgerStatusSyncer />
-      <StyledAppHeader $showFull={showFullStatus} $autoWidth={autoWidthStatus}>
-        <AllRouter />
-      </StyledAppHeader>
+      <AllRouter />
     </StyledApp>
   );
 }